Freeze objects
Modifies all classes to be frozen. This will make attestation of
docments and statements more clear and remove the need to use the
last_updated attributes. If data is modified, though helper functions
added laster on, a new object will be returned with an updated
timestamp. Besides aiding with attestation and timestamps, objects
in this library become hasable, which could be useful to users.

Signed-off-by: Colin Pinnell McAllister <colinmca242@gmail.com>