Remove document timestamp normalization
Removes the model validator that removed timestamp attributes for
statement timestamps that match the document's. Although OpenVEX
supports this, adding successive statements can cause timestamps to not
be updated for statements that have had their timestamp removed. The
complexity of solving this issue to support this aspect of the OpenVEX
spec is not worth it, in my opinion. JSON is a heavy schema as-is and
keeping redundant timestamps in the statement doesn't drastically
increase the size of the JSON.

Signed-off-by: Colin Pinnell McAllister <colinmca242@gmail.com>